update/style: updaeted component locations and styled most components, need to pull current to updata styling
