+++ src/actions/businessActions.js
@@ -25,79 +25,80 @@ export const fetchBusinesses = business => dispatch => {
 	console.log('action business query', business)
 
 	const name = business.name
-	let location
+	//To do!!! set up fetchbusiness to return when user selects business
+	// let location
 
-	// Check for type of location provided - coords or phrase (ie city, state, etc)?
-	if (business.location.latitude && business.location.longitude) {
-		location = `latitude=${business.location.latitude}&longitude=${business.location.longitude}`
-	} else if (business.location) {
-		location = `location=${business.location}`
-	} else {
-		dispatch({
-			type: FETCH_BUSINESS_FAILURE,
-			payload: 'Business location required'
-		})
-	}
-
+	// // Check for type of location provided - coords or phrase (ie city, state, etc)?
+	// if (business.location.latitude && business.location.longitude) {
+	// 	location = `latitude=${business.location.latitude}&longitude=${business.location.longitude}`
+	// } else if (business.location) {
+	// 	location = `location=${business.location}`
+	// } else {
+	// 	dispatch({
+	// 		type: FETCH_BUSINESS_FAILURE,
+	// 		payload: 'Business location required'
+	// 	})
+	// }
+	
+	// axiosWithAuth()
+	// 	.post("/business/search", business)
+	// 		.then(res => {
+	// 			console.log(res)
+	// 		})
+	// 		.catch(err => {
+	// 			console.log(err)
+	// 		})
 	// Dynamically generate endpoint using provided location and name
-	const yelpSearchEndpoint = `https://cors-anywhere.herokuapp.com/https://api.yelp.com/v3/businesses/search?term=${name}&${location}` //I've tried like a million different solutions from Google to get this to work without a 403 and a CORS error, maybe someone else has ideas cause I give up
+	// const yelpSearchEndpoint = `https://cors-anywhere.herokuapp.com/https://api.yelp.com/v3/businesses/search?term=${name}&${location}` //I've tried like a million different solutions from Google to get this to work without a 403 and a CORS error, maybe someone else has ideas cause I give up
 
-	dispatch({ type: FETCH_BUSINESS_START })
-	console.log('Yelp API URL: ', yelpSearchEndpoint)
-	axiosWithYelpAuth()
-		.get(yelpSearchEndpoint)
-		.then(res => {
-			dispatch({
-				type: FETCH_BUSINESS_SUCCESS,
-				payload: res.data.businesses
-			})
-		})
-		.catch(err => {
-			dispatch({
-				type: FETCH_BUSINESS_FAILURE,
-				payload: err
-			})
-		})
+	// dispatch({ type: FETCH_BUSINESS_START })
+	// console.log('Yelp API URL: ', yelpSearchEndpoint)
+	// axiosWithAuth()
+	// 	.post("/business/search", {})
+	// 	.then(res => {
+	// 		dispatch({
+	// 			type: FETCH_BUSINESS_SUCCESS,
+	// 			payload: res.data.businesses
+	// 		})
+	// 	})
+	// 	.catch(err => {
+	// 		dispatch({
+	// 			type: FETCH_BUSINESS_FAILURE,
+	// 			payload: err
+	// 		})
+	// 	})
 }
 
 export const selectBusiness = (
 	previousBusinessInfo,
 	businessInfo
 ) => dispatch => {
-	if (businessInfo) {
-		//only request yelp data if both previous and new info are provided. This is useful while selecting new tabs
-		console.error(
-			'SECOND PARAMETER PROVIDED, not selecting new tab, params: ',
-			previousBusinessInfo,
-			businessInfo
-		)
-		dispatch({ type: SELECT_BUSINESS_START, payload: businessInfo }) //Predict that the selection will be successful by setting currentlySelectedBusiness immediatly, so that the user doesn't need to wait until address/rating info is gotten
-		axiosWithYelpAuth()
-			.get(
-				'https://cors-anywhere.herokuapp.com/https://api.yelp.com/v3/businesses/' +
-					businessInfo.businessId
-			)
-			.then(res => {
-				dispatch({
-					type: SELECT_BUSINESS_SUCCESS,
-					payload: {
-						...businessInfo,
-						address: res.data.location.display_address,
-						reviewCount: res.data.review_count,
-						averageRating: res.data.rating
-					}
-				})
-			})
-			.catch(err => {
-				dispatch({
-					type: SELECT_BUSINESS_FAILURE,
-					payload: previousBusinessInfo //revert back!
-				})
-			})
-	} else {
-		dispatch({ type: SELECT_BUSINESS_SUCCESS, payload: previousBusinessInfo })
-	}
-	console.log('\nAdding business selection to the store...\n')
+	//To DO!! update request since new user not signed in can request
+
+	dispatch({type: SELECT_BUSINESS_START})
+
+	axiosWithAuth()
+		.get(`/business/${businessInfo.business_id}`)
+		.then(res => {
+			const businessData = res.data[0]
+			dispatch({type: SELECT_BUSINESS_SUCCESS, payload: businessData})
+
+// 			address: "2706 E University Dr"
+		// attributes: "{'Caters': 'False', 'OutdoorSeating': 'False', 'BusinessAcceptsCreditCards': 'True', 'BikeParking': 'True', 'HasTV': 'True', 'RestaurantsAttire': "u'casual'", 'RestaurantsDelivery': 'True', 'RestaurantsReservations': 'False', 'RestaurantsTakeOut': 'True', 'Alcohol': "u'none'", 'GoodForKids': 'True', 'RestaurantsPriceRange2': '1', 'BusinessParking': "{'garage': False, 'street': False, 'validated': False, 'lot': False, 'valet': False}", 'Ambience': "{'touristy': False, 'hipster': False, 'romantic': False, 'divey': False, 'intimate': False, 'trendy': False, 'upscale': False, 'classy': False, 'casual': False}", 'WiFi': "u'no'", 'NoiseLevel': "u'quiet'", 'RestaurantsGoodForGroups': 'True', 'GoodForMeal': "{'dessert': False, 'latenight': False, 'lunch': True, 'dinner': True, 'brunch': False, 'breakfast': False}"}"
+		// business_id: "j9bWpCRwpDVfwVT_V85qeA"
+		// business_stars: 2.5
+		// city: "Mesa"
+		// cuisine: "thai"
+		// latitude: 33.4237052
+		// longitude: -111.7728895
+		// name: "Papaya Thai"
+		// review_count: 130
+		// zipcode: "85213"
+		})
+		.catch(err => {
+			console.log(err)
+			dispatch({type: SELECT_BUSINESS_FAILURE, payload: err})
+		})
 }
 
 export const addBusiness = (businessInfo, userID) => dispatch => {
@@ -120,7 +121,7 @@ export const addBusiness = (businessInfo, userID) => dispatch => {
 		city: businessInfo.city,
 		state: businessInfo.state,
 		yelp: {
-			yelp_id: businessInfo.businessId,
+			yelp_id: businessInfo.business_id,
 			url: businessInfo.url,
 			image_url: businessInfo.image_url
 		}
@@ -131,7 +132,7 @@ export const addBusiness = (businessInfo, userID) => dispatch => {
 		backendFormat,
 		businessInfo
 	)
-	//   businessId: "aC1dn3qBFxgk-OYC3hFMgQ"
+	//   business_id: "aC1dn3qBFxgk-OYC3hFMgQ"
 	// businessName: "In The Bowl"
 	// businessImg: "https://s3-media1.fl.yelpcdn.com/bphoto/NpaN9bQ0YsJfI6fEVL5_Qg/o.jpg"
 	// reviewCount: 709
@@ -159,14 +160,14 @@ export const addBusiness = (businessInfo, userID) => dispatch => {
 		})
 }
 
-export const removeBusiness = (businessID, userID) => dispatch => {
+export const removeBusiness = (business_id, userID) => dispatch => {
 	console.log('\removing business from the store...\n')
 	//dispatch({ type: REMOVE_BUSINESS_START, payload: businessInfo });
 	//DELETE /users/:id/business/:business_id
 	dispatch({ type: REMOVE_BUSINESS_START })
 	//endpoint
 	axiosWithAuth()
-		.delete(`/users/${userID}/business/${businessID}`)
+		.delete(`/users/${userID}/business/${business_id}`)
 		.then(res => {
 			dispatch({
 				type: REMOVE_BUSINESS_SUCCESS,
